During Screen's update frame cycle, there is a calculation which is used to determine how long to wait before updating and displaying the next frame:
delay = (1000 ms /FPS) - (time taken to prepare and display the last frame)
(where FPS = 25)

If delay is negative boost::thread::sleep(..) waits indefinitely.  The effect in Asgard is that the updating of Screen animations stops.  I've found that delay is most often negative after maximizing a window which has been previously minimized.  The fix was to set delay = 1 in the event that the calculation of delay <= 0.  Later, we should consider adjusting the FPS down if the delay is consistently near 0.

Reference: https://svn.boost.org/trac/boost/ticket/3178